Re-organized modules so the API uncoupled from the REST Client.
