+++ pom.xml
@@ -74,6 +74,7 @@
 
   <modules>
     <module>common</module>
+    <module>api</module>
     <module>server</module>
     <module>client</module>
     <module>bundle</module>
